refactor: made directories for page sections test: added tests for all sections
