+++ .eslintrc.json
@@ -20,7 +20,8 @@
             "**/*.spec.jsx"
           ],
           "env": {
-            "jest": true
+            "jest": true,
+            "browser": true
           }
         }
       ]
