+++ src/components/sections/footer-section/footer-section.test.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+import FooterSection from './footer-section';
+
+describe('Footer Section Test Suite', () => {
+  it('renders correctly', () => {
+    const tree = renderer.create(<FooterSection />).toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+});
