+++ src/components/sections/title-section/title-section.test.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+import TitleSection from './title-section';
+
+describe('Title Section Test Suite', () => {
+  it('renders correctly', () => {
+    const tree = renderer.create(<TitleSection />).toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+});
