+++ src/pages/index.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import Seo from '../components/seo/seo';
 import ContactSection from '../components/sections/contact-section';
-import AboutSection from '../components/sections/about-section';
+import AboutSection from '../components/sections/about-section/about-section';
 import TitleSection from '../components/sections/title-section';
-import Navbar from '../components/sections/navbar/navbar';
+import Navbar from '../components/navbar/navbar';
 import Footer from '../components/sections/footer-section';
 import '../components/styles.css';
